requirements.txt
Flask==3.0.0
groq==0.4.1
python-dotenv==1.0.0
